applied Roberts comments about placement + fixed bug import, where DBS_nr was filepath instead of filename
