+++ src/__init__.py
@@ -1,5 +1,6 @@
 from flask import Flask
 from flask_pymongo import PyMongo
+from flask_admin import Admin
 #from flask_mongoengine import MongoEngine
 from flask_login import LoginManager
 
@@ -22,5 +23,9 @@ login_manager.login_message_category = "warning"
 # Flask BCrypt will be used to salt the user password
 # flask_bcrypt = Bcrypt(app)
 
+# Create admin
+admin = Admin(app, name='VUSualizer admin', template_mode='bootstrap3')
+
+
 from . import *
-from .views import main, auth, api
+from .views import main, auth, api, admin_view
