+++ src/templates/admin/index.html
@@ -14,11 +14,9 @@
             </p>
             {% endif %}
             {% if not current_user.is_authenticated %}
-            <p>You can register as a regular user, or log in as a superuser with the following credentials:
-            <ul>
-                <li>email: <b>admin</b></li>
-                <li>password: <b>admin</b></li>
-            </ul>
+            <p>You are currently logged in as {{ current_user.email }}, with role {{ current_user.role }}
+            <p>You are not an Admin user and can therefore not access the Admin views. For more information contact f.e.degruyter@umcutrecht.nl
+
             <p>
                 <a class="btn btn-primary" href="{{ url_for('login') }}">login</a> <a class="btn btn-default" href="{{ url_for('register') }}">register</a>
             </p>
